refactor(test): split backend test monolith into suite files
